+++ modules/Accounting/Incomes.php
@@ -185,7 +185,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 		DrawHeader( '', SubmitButton() );
 
-		$options = [];
+		$options = [ 'valign-middle' => true ];
 	}
 	else
 	{